\name{c}
\alias{c,feature-method}
\title{
    Combine Multiple Features to Give a List of Features
}
\description{
  Combine multiple \code{feature} objects to give a list of feature objects
}

\usage{
\S4method{c}{feature}(x, ..., recursive = FALSE)
}
\arguments{
    \item{x}{
      An object of class \code{feature}
    }
    \item{...}{
      One or more \code{feature} objects
    }
    \item{recursive}{ Ignored }
}
\author{Rajarshi Guha \email{rajarshi.guha@gmail.com}}
\keyword{logic}



