\name{length}
\alias{length}
\alias{length,fingerprint-method}
\title{
  Fingerprint Bit Length
}
\description{
  Returns the length of the fingerprint. That is, this is the length of
  the entire bit string and not simply the number of bits that are on.
}
\usage{
  \S4method{length}{fingerprint}(x)
}
\arguments{
    \item{x}{
      An object of class \code{fingerprint}
    }
}
\value{
The length of the bit string
}
\keyword{logic}
\keyword{methods}
\author{Rajarshi Guha \email{rguha@indiana.edu}}



