\name{show}
\alias{show,fingerprint-method}
\alias{show,featvec-method}
\alias{show,feature-method}
\title{
    String Representation of a Fingerprint or Feature
}
\description{
Simply summarize the fingerprint or feature
}

\usage{
\S4method{show}{fingerprint}(object)
\S4method{show}{featvec}(object)
\S4method{show}{feature}(object)
}
\arguments{
    \item{object}{
      An object of class \code{fingerprint}, \code{featvec} or \code{feature}
    }
}
\author{Rajarshi Guha \email{rajarshi.guha@gmail.com}}
\keyword{logic}



