\name{show}
\alias{show,fingerprint-method}
\title{
    String Representation of a Fingerprint
}
\description{
Simply summarize the fingerprint.
}

\usage{
\S4method{show}{fingerprint}(object)
}
\arguments{
    \item{object}{
      An object of class \code{fingerprint}
    }
}
\author{Rajarshi Guha \email{rguha@indiana.edu}}
\keyword{logic}



