\name{show}
\alias{show,fingerprint-method}
\alias{show,featvec-method}
\title{
    String Representation of a Fingerprint
}
\description{
Simply summarize the fingerprint.
}

\usage{
\S4method{show}{fingerprint}(object)
}
\arguments{
    \item{object}{
      An object of class \code{fingerprint} or \code{featvec}
    }
}
\author{Rajarshi Guha \email{rajarshi.guha@gmail.com}}
\keyword{logic}



